Add more pages
- Add sign in page
- Add sign up page
- Add Learning plan components
- Add event listener to handle account authentications

This commit adds sign in and signup functionality to the platform, althought not fully implemented, basic functionality has been tested and is working. The next task will to respond to successful authentication of user by creating the user profile.

The learning plan page is not implemented properly, but what we have now is a simple model that works for basic requirements.